requirements.txt
flask==3.0.0
sentence-transformers==5.2.3
pypdf==4.0.2
python-dotenv==1.0.0
langchain==1.2.11
langchain-core==1.2.18
langchain-community==0.4.1
langchain-openai==1.1.11
langchain-groq==1.1.2
pinecone==7.3.0
langchain-pinecone==0.2.13

